Improve solution unpacking
The function got renamed to findSolutionDir as it is specifically made
to find a solution (not a general directory matching just a condition).
Some short documentation was added to the function as well.
Also, it does now stop descending if either the stop condition is true
or the directory is not a wrapper directory.